% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_pdflatex}
\alias{check_pdflatex}
\title{Check for LaTeX}
\usage{
check_pdflatex()
}
\value{
Warning side-effects informing users about missing LaTeX
and specifically PdfLaTeX.
}
\description{
Check for LaTeX
}
\keyword{internal}
