% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\alias{FILE_HOSP_OVERFLOW_RISK}
\alias{FILE_DEATHS_BY_AGE}
\alias{FILE_COST_BY_RESPONSE}
\alias{FILE_ECON_COST_BREAKDOWN}
\title{Package constants}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
FILE_HOSP_OVERFLOW_RISK

FILE_DEATHS_BY_AGE

FILE_COST_BY_RESPONSE

FILE_ECON_COST_BREAKDOWN
}
\description{
Package constants
}
\examples{

FILE_HOSP_OVERFLOW_RISK

FILE_DEATHS_BY_AGE

FILE_COST_BY_RESPONSE

FILE_ECON_COST_BREAKDOWN

}
\keyword{constants}
