% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{df_cols_with_commas}
\alias{df_cols_with_commas}
\title{Add commas to numeric columns}
\usage{
df_cols_with_commas(df)
}
\arguments{
\item{df}{A data.frame with columns matching \code{"pctl"}.}
}
\value{
A data.frame with columns matching \code{"pctl"} converted to type
character and with commas added to make reading large numbers easier.
Numbers are rounded to the nearest ones-place, or to the nearest 10th if
less than 1.
}
\description{
Add commas to numeric columns
}
