% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_themes.R
\name{figure_themes}
\alias{figure_themes}
\alias{theme_eppi}
\alias{theme_explainer}
\title{Theme for handout figures}
\usage{
theme_eppi()

theme_explainer()
}
\value{
A \code{ggplot2} theme function that can be appended to a ggplot object.
}
\description{
Theme for handout figures
}
