% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{has_pdflatex}
\alias{has_pdflatex}
\title{Check for LaTeX}
\usage{
has_pdflatex()
}
\value{
A single logical for whether LaTeX (and specifically PdfLaTeX) is
available. Also has warning side-effects informing users about missing LaTeX
and PdfLaTeX.
}
\description{
Check for LaTeX
}
