% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_deaths_by_age_table}
\alias{make_deaths_by_age_table}
\title{Make deaths by age table}
\usage{
make_deaths_by_age_table(tables_out)
}
\arguments{
\item{tables_out}{Location to table outputs.}
}
\value{
A \code{knitr::kable()} table.
}
\description{
Make deaths by age table
}
