% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_hcap_breaches_table}
\alias{make_hcap_breaches_table}
\title{Make hospital capacity breaches table}
\usage{
make_hcap_breaches_table(tables_out)
}
\arguments{
\item{tables_out}{Location to table outputs.}
}
\value{
A \code{knitr::kable()} table.
}
\description{
Make hospital capacity breaches table
}
