% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{NAMES_PRECANNED_NPIS}
\alias{NAMES_PRECANNED_NPIS}
\title{name constants}
\format{
An object of class \code{character} of length 4.
}
\usage{
NAMES_PRECANNED_NPIS
}
\description{
name constants
}
\keyword{constants}
