% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{theme_eppi}
\alias{theme_eppi}
\title{Theme for handout figures}
\usage{
theme_eppi()
}
\value{
A \code{ggplot2} theme function that can be appended to a ggplot object.
}
\description{
Theme for handout figures
}
