% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{constants}
\alias{constants}
\alias{FILE_HOSP_OVERFLOW_RISK}
\alias{FILE_DEATHS_BY_AGE}
\alias{FILE_COST_BY_RESPONSE}
\alias{FILE_ECON_COST_BREAKDOWN}
\alias{NAMES_PRECANNED_NPIS}
\title{Package constants}
\usage{
FILE_HOSP_OVERFLOW_RISK

FILE_DEATHS_BY_AGE

FILE_COST_BY_RESPONSE

FILE_ECON_COST_BREAKDOWN

NAMES_PRECANNED_NPIS
}
\description{
Package constants
}
\examples{

FILE_HOSP_OVERFLOW_RISK

FILE_DEATHS_BY_AGE

FILE_COST_BY_RESPONSE

FILE_ECON_COST_BREAKDOWN

}
\keyword{constants}
