% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daedalus.workshop-package.R
\docType{package}
\name{daedalus.workshop-package}
\alias{daedalus.workshop}
\alias{daedalus.workshop-package}
\title{daedalus.workshop: Generate Pandemic Response Workshop Materials Using Daedalus}
\description{
Automate the generation of pandemic response workshop materials using the Daedalus family of packages.
}
\author{
\strong{Maintainer}: Pratik Gupte \email{pratik.gupte@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5294-7819}{ORCID})

Other contributors:
\itemize{
  \item Abdul Latif Jameel Institute for Disease and Emergency Analytics [funder]
  \item Imperial College of Science, Technology and Medicine [copyright holder, funder]
}

}
\keyword{internal}
