% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_domain_costs_table}
\alias{make_domain_costs_table}
\title{Make domain costs table}
\usage{
make_domain_costs_table(tables_out)
}
\arguments{
\item{tables_out}{Location to table outputs.}
}
\value{
A \code{knitr::kable()} table.
}
\description{
Make domain costs table
}
