% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_econ_cost_table}
\alias{make_econ_cost_table}
\title{Make economic costs breakdown table}
\usage{
make_econ_cost_table(tables_out)
}
\arguments{
\item{tables_out}{Location to table outputs.}
}
\value{
A \code{knitr::kable()} table.
}
\description{
Make economic costs breakdown table
}
