% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_materials}
\alias{make_materials}
\title{Generate all workshop materials}
\usage{
make_materials(
  country = "GBR",
  disease = "sars_cov_1",
  r0 = 3,
  t0 = 30,
  t_end = 100
)
}
\description{
Generates all workshop materials in the current working directory.
}
