% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_materials}
\alias{make_materials}
\title{Generate all workshop materials}
\usage{
make_materials(
  country = "GBR",
  disease = "sars_cov_1",
  t0 = 30,
  horizon = 100,
  n_samples = 10,
  workshop_name = "Pandemic Response Workshop",
  date = as.character(Sys.Date()),
  currency = "$",
  render = TRUE
)
}
\arguments{
\item{country}{The country name or three-letter ISO code.}

\item{disease}{The disease name. Must be one of
\link[daedalus.data:epidemic_data]{daedalus.data::epidemic_names}.}

\item{t0}{The current time-point. Defaults to 30 days.}

\item{horizon}{The time horizon for projections. Defaults to 100 days.}

\item{n_samples}{The number of samples for parameter uncertainty. Default 10.}

\item{workshop_name}{The workshop name. Intended to be used as a sub-title.}

\item{date}{The workshop date as a string. Defaults to the current date.}

\item{currency}{A string for the currency name or symbol to be used. A symbol
is preferred as the final use is in the form \code{"<symbol>M"} to indicate
millions in \code{currency}.}

\item{render}{Whether the drafted Rmarkdown document should be rendered.}
}
\value{
Nothing; called only for side-effects of drafting and rendering
Rmarkdown documents for real-time pandemic response workshop materials.
}
\description{
Generates all workshop materials in the current working directory.
}
