% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_materials}
\alias{make_materials}
\title{Generate all workshop materials}
\usage{
make_materials(
  country = "GBR",
  disease = "sars_cov_1",
  r0 = 3,
  t0 = 30,
  horizon = 100,
  n_samples = 10,
  render = TRUE
)
}
\arguments{
\item{country}{description}

\item{disease}{}

\item{r0}{}

\item{t0}{}

\item{horizon}{}

\item{n_samples}{}

\item{render}{}
}
\value{
Nothing; called only for side-effects of drafting and rendering
Rmarkdown documents for real-time pandemic response workshop materials.
}
\description{
Generates all workshop materials in the current working directory.
}
