% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_sector_table}
\alias{make_sector_table}
\title{Make table of sector GVA and contacts}
\usage{
make_sector_table(country)
}
\arguments{
\item{country}{A country name, or a type that can be coerced to a
\verb{<daedalus_country>}.}
}
\value{
A \code{knitr::kable()} table.
}
\description{
Make table of sector GVA and contacts
}
